Also mutate containers if they don't have secret references
If the container is explicitely configured to be mutated, mutate
it also if there are no secret references in the environment of the
pod spec. This allows for other ways to use secrethub run. For
example, you could supply a secrethub.env file over a volume or
you could set secret reference environment variables in some other
way.